moved source xml elements and attribute enums to util
