+++ src/org/opensha2/eq/model/AreaParser.java
@@ -4,15 +4,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.util.logging.Level.FINE;
 
-import static org.opensha2.eq.model.SourceAttribute.FOCAL_MECH_MAP;
-import static org.opensha2.eq.model.SourceAttribute.ID;
-import static org.opensha2.eq.model.SourceAttribute.MAG_DEPTH_MAP;
-import static org.opensha2.eq.model.SourceAttribute.MAX_DEPTH;
-import static org.opensha2.eq.model.SourceAttribute.NAME;
-import static org.opensha2.eq.model.SourceAttribute.RUPTURE_SCALING;
-import static org.opensha2.eq.model.SourceAttribute.STRIKE;
-import static org.opensha2.eq.model.SourceAttribute.TYPE;
-import static org.opensha2.eq.model.SourceAttribute.WEIGHT;
 import static org.opensha2.eq.model.SourceType.AREA;
 import static org.opensha2.util.Parsing.readDouble;
 import static org.opensha2.util.Parsing.readEnum;
@@ -20,6 +11,15 @@ import static org.opensha2.util.Parsing.readInt;
 import static org.opensha2.util.Parsing.readString;
 import static org.opensha2.util.Parsing.stringToEnumWeightMap;
 import static org.opensha2.util.Parsing.stringToValueValueWeightMap;
+import static org.opensha2.util.SourceAttribute.FOCAL_MECH_MAP;
+import static org.opensha2.util.SourceAttribute.ID;
+import static org.opensha2.util.SourceAttribute.MAG_DEPTH_MAP;
+import static org.opensha2.util.SourceAttribute.MAX_DEPTH;
+import static org.opensha2.util.SourceAttribute.NAME;
+import static org.opensha2.util.SourceAttribute.RUPTURE_SCALING;
+import static org.opensha2.util.SourceAttribute.STRIKE;
+import static org.opensha2.util.SourceAttribute.TYPE;
+import static org.opensha2.util.SourceAttribute.WEIGHT;
 
 import org.opensha2.data.Data;
 import org.opensha2.eq.fault.FocalMech;
@@ -28,6 +28,7 @@ import org.opensha2.geo.LocationList;
 import org.opensha2.mfd.IncrementalMfd;
 import org.opensha2.mfd.MfdType;
 import org.opensha2.mfd.Mfds;
+import org.opensha2.util.SourceElement;
 
 import org.xml.sax.Attributes;
 import org.xml.sax.Locator;
